chore gradle-wrapper.propertiesにorg.gradle.warning.mode=allを追加
