+++ src/main/java/com/sixassignment/helloworld/HelloworldApplication.java
@@ -1,11 +1,10 @@
-package com.six_assignment.helloworld;
+package com.sixassignment.helloworld;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class HelloworldApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(HelloworldApplication.class, args);
     }
