+++ src/test/java/com/sixassignment/helloworld/HelloworldApplicationTests.java
@@ -1,4 +1,4 @@
-package com.six_assignment.helloworld;
+package com.sixassignment.helloworld;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class HelloworldApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
